# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPets",
        "description": "bark",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "All friends",
                "description": "Get pet friends.",
                "operationId": "friendsAll",
                "responses": {
                    "200": {
                        "description": "Friends of pet.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "gender": {
                                                "type": "string"
                                            },
                                            "size": {
                                                "type": "string"
                                            },
                                            "primary_breed": {
                                                "type": "string"
                                            },
                                            "secondary_breed": {
                                                "type": "string"
                                            },
                                            "age": {
                                                "type": "integer"
                                            },
                                            "profile_picture": {
                                                "type": "string"
                                            },
                                            "friendliness": {
                                                "type": "integer"
                                            },
                                            "activity_level": {
                                                "type": "integer"
                                            },
                                            "noise_level": {
                                                "type": "integer"
                                            },
                                            "odebience_level": {
                                                "type": "integer"
                                            },
                                            "fetchability": {
                                                "type": "integer"
                                            },
                                            "swimability": {
                                                "type": "integer"
                                            },
                                            "city": {
                                                "type": "string"
                                            },
                                            "state": {
                                                "type": "string"
                                            },
                                            "like": {
                                                "type": "string"
                                            },
                                            "dislike": {
                                                "type": "string"
                                            },
                                            "favorite_toys": {
                                                "type": "string"
                                            },
                                            "fears": {
                                                "type": "string"
                                            },
                                            "favorite_places": {
                                                "type": "string"
                                            },
                                            "spayed": {
                                                "type": "string"
                                            },
                                            "birthday": {
                                                "type": "date"
                                            },
                                            "pictures": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "picture": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated error or registration error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated.|Sign-Up steps not done."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}